Handle awkward edge TLS case in old Node versions too
This is reproducible in Mockttp on Node v16 & v18, when a socket is
intentionally reset & destroyed, the TLS socket on top is not marked
as such, and so this never resolved. This fix ensures that even TLS
sockets which don't correctly get marked as destroyed are still
considered as closed once their underlying connection is closed.